feat: Add company color code to registration process
This commit modifies the `registerCompanyAndUser` function in the `company.controller.js` file to include the company color code in the registration process. If a `colorCode` is provided, it is assigned to the `company_info.companyColorCode` property. If no `colorCode` is provided, a default color code is assigned. Additionally, the `companyAdmin` field is updated with the user's ID and the company document is saved.