```text feat: Add support for extra query parameters in getAllUsers endpoint
This commit modifies the getAllUsers endpoint in the user.controller.js file to support additional query parameters. The code now checks for the presence of the "extra" query parameter and includes it in the data object for filtering users. This enhancement allows for more flexible filtering options when retrieving users.