refactor: Update UserGroup controller to populate userCount field
This commit modifies the UserGroup controller to populate the userCount field when retrieving all groups. The code change adds a query to count the number of users associated with each group and assigns the count to the userCount field of the UserGroup object. This enhancement improves the completeness of the data by providing the user count for each group.