refactor: Fix user group access level check in user.controller.js
The code change in user.controller.js fixes the user group access level check in the registerNewUser function. The condition is modified to correctly access the accessLevel property of the user group from the request object. This fix ensures that only admin or manager users can add new users, improving the security and access control of the application.