refactor: Update auth.middleware.js to populate userGroup field
This commit modifies the auth.middleware.js file to populate the userGroup field when retrieving the user. The code change adds the `populate("userGroup")` method to the `User.findById` query, ensuring that the user's associated user group is included in the response. This enhancement improves the completeness of the data by providing the user group information along with the user object.