refactor: Update task controller to include team leader details
The code changes in `task.controller.js` add a lookup stage to retrieve the team leader details for each task. This enhancement populates the `teamLeaderDetails` field in the task object, providing additional information about the team leader assigned to the task. This improves the completeness of the data and enhances the functionality of the application.