refactor: Add start-local script to package.json for local development
The code changes in `package.json` add a new script called `start-local` to facilitate local development. This script allows developers to start the application locally using the command `npm run start-local`. By adding this script, it provides a convenient way to run the application without the need for additional configuration. This refactor improves the developer experience and streamlines the local development process.